feat: export `Sandbox` as standalone output component (#309)
* feat: export `Sandbox`

* docs: sandbox case
