+++ src/index.ts
@@ -1,5 +1,7 @@
 export { default as Repl } from './Repl.vue'
 export { default as Preview } from './output/Preview.vue'
+export { default as Sandbox } from './output/Sandbox.vue'
+export type { SandboxProps } from './output/Sandbox.vue'
 export {
   useStore,
   File,
